Split lexer number literal tests
