+++ docs/COMPLETION_AUDIT.md
@@ -2720,6 +2720,10 @@ and do not assume Phase 4 is ready without evidence.
 - Lexer string literal and interpolation tests now live in
   `src/lexer/tests/string_literals.rs`, preserving literal escape and
   interpolation coverage while keeping the parent lexer test module smaller.
+- Lexer numeric literal and range-token tests now live in
+  `src/lexer/tests/number_literals.rs`, preserving integer, float, prefixed
+  integer, and float-vs-range token coverage while keeping the parent lexer
+  test module smaller.
 - Lexer whitespace and comment skipping now lives in `src/lexer/whitespace.rs`,
   preserving newline-token and string-interpolation comment coverage while
   keeping token dispatch smaller.
